requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
yfinance==0.2.36
pytz==2024.1

